client/requirements.txt
# 客户端同步工具依赖
# 核心HTTP和文件操作
requests>=2.28.0
PyYAML>=6.0

# 可选：更好的进度显示
tqdm>=4.64.0

# 可选：日志增强
colorlog>=6.7.0